requirements.txt
appdirs==1.4.4
bcrypt==3.1.7
cffi==1.15.1
cryptography==2.8
Fabric3==1.14.post1
paramiko==2.12.0
pycparser==2.21
PyNaCl==1.3.0
pyparsing==3.0.9
six==1.16.0
